Handle anonymous types for incremental analysis.
CIL creates type definitions for anonymous types (structs and unions).
The names for these types contain a hashvalue of the "context" of these
type definitions. When the formatting of the definition of an anonymous
type changes, these type names may change, too.
Thus, a changed formatting may cause a type to be considered different,
as well as all variables of this type or any type that cotains this
type as a field. We want to avoid this.

An anonymous type that is structurally the same as an old anonymous type
and is used in at least one place that is semantically related to an use
of the old anonymous type, will be renamed.

The current implementation may be over-eager in doing renamings.