Fix def_exc
`Exluded is equipped with a range the represents the smallest C Data-type that can contain both `Definite values when they are joined. This is then used when calling minimal and maximal. However, this stayed static when doing computations such as addition, making the values returned from maximal and minimal unsound.