Add testcase showing multiplication (w/o 0) is not injective.
This shows a problem in the treatment of operations:
    (Excluded (s, r)) op (Definite y)

Excluded(s op y, _) is in general not a correct result value here.
(Where s op y := {x op y | x in s})
This is because there might be an z in gamma(Excluded (s, r)) such that
    z op y (mod 2^32)
is in (s op y).

Some overflow checking is needed.