Add signature IntDomain.Z and IntDomLifter(I: S): Z
Some preliminary work to store the ikind for abstract integer values.
We want to have the ikind available for all abstract integer values, when
we perform operations on them, but we don't really want to store the
ikind redundantly in each of the integer domain implementations.

We introduce a IntDomLifter, that takes a integer domain implementation
and adds the ikind information to it. In the future, this
IntDomLifter should pass the ikind to the domain operations that need it.
This is the reason why we need a separate signature Z for instantiations
of the IntDomLifter which is different from the signature S.
In the signature Z only the creation of abstract value expects an ikind,
but - unlike S - the integer operations itself will not require it.

Needs to be reworked inorder not to fail test cases.